API

SHTns.ForRotations

SHTns.FourPi

SHTns.Gauss

SHTns.GaussFly

SHTns.Orthonormal

SHTns.QuickInit

SHTns.RegDCT

SHTns.RegFast

SHTns.RegPoles

SHTns.SHTnsCfg

SHTns.Schmidt

SHTns.LM

SHTns.LM_cplx

SHTns.grid

SHTns.k2lm

SHTns.SHTnsCfg — Type

mutable struct SHTnsCfg{N<:SHTnsNorm, T<:SHTnsType}

Configuration of spherical harmonic transform.

SHTns.LM — Method

LM(cfg, l, m)

Returns index corresponding to l and m in real spherical harmonic expansion.

SHTns.LM_cplx — Method

LM_cplx(cfg, l, m)

Returns index corresponding to l and m in complex spherical harmonic expansion.

SHTns.grid — Method

grid(cfg::SHTnsCfg)

Returns latitudes lat::Vector{Float64} and longitudes lon::Vector{Float64}.

SHTns.ForRotations — Type

struct SHTns.ForRotations <: SHTnsNorm

SHTns.FourPi — Type

struct SHTns.FourPi <: SHTnsNorm

SHTns.Gauss — Type

struct SHTns.Gauss <: SHTnsType

SHTns.GaussFly — Type

struct SHTns.GaussFly <: SHTnsType

SHTns.Orthonormal — Type

struct SHTns.Orthonormal <: SHTnsNorm

SHTns.QuickInit — Type

struct SHTns.QuickInit <: SHTnsType

SHTns.RegDCT — Type

struct SHTns.RegDCT <: SHTnsType

SHTns.RegFast — Type

struct SHTns.RegFast <: SHTnsType

SHTns.RegPoles — Type

struct SHTns.RegPoles <: SHTnsType

SHTns.Schmidt — Type

struct SHTns.Schmidt <: SHTnsNorm

SHTns.k2lm — Method

k2lm(cfg, idx)

Returns the l and m corresponding to the given index idx in complex spherical harmonic expansion.