API

SHTns.ForRotations

SHTns.FourPi

SHTns.Gauss

SHTns.GaussFly

SHTns.Orthonormal

SHTns.QuickInit

SHTns.RegDCT

SHTns.RegFast

SHTns.RegPoles

SHTns.SHTnsCfg

SHTns.Schmidt

SHTns.LM

SHTns.LM

SHTns.gauss_weights

SHTns.grid

SHTns.synth

SHTns.synth

SHTns.synth

SHTns.ForRotations — Type

struct SHTns.ForRotations <: SHTnsNorm

SHTns.FourPi — Type

struct SHTns.FourPi <: SHTnsNorm

SHTns.Gauss — Type

struct SHTns.Gauss <: SHTnsType

SHTns.GaussFly — Type

struct SHTns.GaussFly <: SHTnsType

SHTns.Orthonormal — Type

struct SHTns.Orthonormal <: SHTnsNorm

SHTns.QuickInit — Type

struct SHTns.QuickInit <: SHTnsType

SHTns.RegDCT — Type

struct SHTns.RegDCT <: SHTnsType

SHTns.RegFast — Type

struct SHTns.RegFast <: SHTnsType

SHTns.RegPoles — Type

struct SHTns.RegPoles <: SHTnsType

SHTns.SHTnsCfg — Type

mutable struct SHTnsCfg{N<:SHTnsNorm, T<:SHTnsType}

Configuration of spherical harmonic transform.

SHTns.Schmidt — Type

struct SHTns.Schmidt <: SHTnsNorm

SHTns.LM — Method

LM(cfg::SHTnsCfg{Complex,T,N}, l, m)

Returns index corresponding to l and m in complex spherical harmonic expansion.

SHTns.LM — Method

LM(cfg::SHTnsCfg{Real,T,N}, l, m)

Returns index corresponding to l and m in real spherical harmonic expansion.

SHTns.gauss_weights — Method

gauss_weights(cfg::SHTnsCfg)

Returns Gauss quadrature weights of length cfg.nlat (shtns_gauss_wts only returns half of the symmetric weights).

SHTns.grid — Method

grid(cfg::SHTnsCfg)

Returns latitudes lat::Vector{Float64} and longitudes lon::Vector{Float64}.

SHTns.synth — Method

synth(cfg::SHTnsCfg, qlm, slm, tlm)

DOCSTRING

Arguments:

cfg: DESCRIPTION

qlm: DESCRIPTION

slm: DESCRIPTION

tlm: DESCRIPTION

SHTns.synth — Method

synth(cfg::SHTnsCfg, slm, tlm)

DOCSTRING

Arguments:

cfg: DESCRIPTION

slm: DESCRIPTION

tlm: DESCRIPTION

SHTns.synth — Method

synth(cfg::SHTnsCfg, qlm)

DOCSTRING